Specific email handling utils fily.
