+++ src/django/VLE/views/update.py
@@ -9,6 +9,7 @@ from rest_framework.parsers import JSONParser
 import VLE.views.responses as responses
 import VLE.serializers as serialize
 import VLE.utils.generic_utils as utils
+import VLE.utils.email_handling as email_handling
 import VLE.permissions as permissions
 import VLE.factory as factory
 import VLE.validators as validators
@@ -828,7 +829,7 @@ def forgot_password(request):
     if not user:
         return responses.bad_request('No user found with that username or email.')
 
-    utils.send_password_recovery_link(user)
+    email_handling.send_password_recovery_link(user)
 
     return responses.success(description='An email was sent to %s, please follow the email for instructions.'
                              % user.email)
@@ -911,7 +912,7 @@ def request_email_verification(request):
     if user.verified_email:
         return responses.bad_request(description='Email address already verified.')
 
-    utils.send_email_verification_link(user)
+    email_handling.send_email_verification_link(user)
 
     return responses.success(description='An email was sent to %s, please follow the email for instructions.'
                              % user.email)
